Move chat to vetsai url, update auth
