+++ src/app/layout.tsx
@@ -5,10 +5,12 @@ export const metadata = {
     description: "Generated by Next.js",
 };
 
-export default function RootLayout({ children }: { children: React.ReactNode }) {
+const RootLayout = ({ children }: { children: React.ReactNode }) => {
     return (
         <html lang="en">
             <body>{children}</body>
         </html>
     );
-}
+};
+
+export default RootLayout;
